Update projects to work with local dev sql
Sql server in docker for dev uses self signed cert. To be able to use
samples with the container needed to relax connection strings.  These
connection strings idealy should not be hard coded into the code and
should be able to adjust all settings by specifing directly.  Will need
to make an update to all the code bases for this change.